refactor(#18): Handle process launch using a kernel process
This change will make the kernel handle the launch of processes using a
kernel process that is launched at the start of the simulation and
launches processes when it receives `Launch` messages.
